// src/models/check-type.ts
export enum CheckType {
  Chest = 'icons/chest.png',
  Skulltula = 'icons/skulltula.png',
  GossipStone = 'icons/gossip-stone.png',
  Scrub = 'icons/rupee.png',
  NPC = 'icons/gift.png',
  Game = 'icons/game.png',
  Song = 'icons/song.png',
  HeartPiece = 'icons/heart.png',
  Key = 'icons/key.png',
}